backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.0

# Google AI
google-generativeai==0.7.2
Pillow==10.3.0

# ElevenLabs
httpx==0.27.0

# MongoDB
motor==3.4.0

# Snowflake
snowflake-connector-python==3.10.1

# WebSockets
websockets==12.0
